Remove customroles
