+++ internal/provider/provider.go
@@ -14,7 +14,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/singlestore-labs/singlestore-go/management"
 	"github.com/singlestore-labs/terraform-provider-singlestoredb/internal/provider/config"
-	"github.com/singlestore-labs/terraform-provider-singlestoredb/internal/provider/customroles"
 	"github.com/singlestore-labs/terraform-provider-singlestoredb/internal/provider/flow"
 	"github.com/singlestore-labs/terraform-provider-singlestoredb/internal/provider/invitations"
 	"github.com/singlestore-labs/terraform-provider-singlestoredb/internal/provider/privateconnections"
@@ -196,7 +195,7 @@ func (p *singlestoreProvider) Resources(_ context.Context) []func() resource.Res
 		roles.NewUserRolesGrantResource,
 		roles.NewTeamRoleGrantResource,
 		roles.NewTeamRolesGrantResource,
-		customroles.NewResource,
+		roles.NewRoleResource,
 		flow.NewResource,
 	}
 }
