+++ internal/provider/roles/role_resource_test.go
@@ -1,4 +1,4 @@
-package customroles_test
+package roles_test
 
 import (
 	"encoding/json"
